Clean up gammapy.image tests
